fix(studio): split color grading inspector files
